[rename] 오타 패키지 명 수정
