+++ src/main/java/com/october/back/security/oauth2/service/CustomOAuth2User.java
@@ -1,4 +1,4 @@
-package com.october.back.security.oauth2.serivce;
+package com.october.back.security.oauth2.service;
 
 import com.october.back.user.entity.dto.UserDto;
 import org.springframework.security.core.GrantedAuthority;
