+++ src/main/java/com/october/back/security/oauth2/service/dto/OAuth2ResponseDto.java
@@ -1,4 +1,4 @@
-package com.october.back.security.oauth2.serivce.dto;
+package com.october.back.security.oauth2.service.dto;
 
 public interface OAuth2ResponseDto {
